implements suggestions
